refactor command main and package up commands
